requirements.txt
yt-dlp>=2023.10.0
pillow